The Player Controller is now a Finite State Machine!
I made the player controller a state machine, so animating the player will now be soooooooooooooooooooooooooooooooooooooooooooooooo much easier, since the logic is already done for the states. I also added an empty core, which for some reason, if you are already touching it then you dash, you don't go into it, I'll figure it out later.